Moved CodableFeedStore to it own file in production
